auth/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic[email]
bcrypt
pyjwt
pydantic-settings 

